+++ JSports/site/jsports.php
@@ -2,5 +2,3 @@
 
 JHtml::_('jquery.framework');
 
-echo "hello chris";
-exit;